requirements.txt
pyenchant
numpy
scipy
scikit-learn
